TST/REF: Consolidate tests in apply.test_invalid_arg (#40092)
